drop 1.9.3 add newer jruby and rubinius
Since 1.9.3 is EOLed and fails on travis, I think it makes sense to drop it, unless there's a need to continue support. I also added 2.3, the latest and head so that there is coverage for the latest ruby stuff. I also added allowed failures on head, as that's a moving target.